fix(sdk): rename `mod core` to avoid shadowing std::core
Renamed:
- sdk/src/structures/core.rs → sdk/src/structures/game_world.rs
- sdk/src/addresses/functions.rs: `pub mod core` → `pub mod engine`

The name `core` shadows Rust's std::core crate, which can cause
subtle compilation errors when using core::mem, core::ptr inside
these modules.

Updated all internal references accordingly.
