+++ sdk/src/addresses/functions.rs
@@ -7,10 +7,10 @@
 //! - R9  = 4-й аргумент или XMM3
 
 // ═══════════════════════════════════════════════════════════════════════════
-//  Core System
+//  Engine System
 // ═══════════════════════════════════════════════════════════════════════════
 
-pub mod core {
+pub mod engine {
     /// Точка входа игры.
     /// IDA: `0x1412CCDC0`
     pub const MAIN_ENTRY_POINT: usize = 0x12C_CDC0;
