+++ sdk/src/structures/game_world.rs
